requirements.txt
faiss-cpu==1.13.2
fastapi==0.127.0
numpy==2.2.6
opencv-python==4.12.0.88
openvino==2025.4.1
pillow==12.0.0
python-multipart==0.0.21
ultralytics==8.3.241
uvicorn==0.40.0
websockets==15.0.1